Move classes to components package
