+++ html/attributes_test.go
@@ -9,29 +9,6 @@ import (
 	. "github.com/maragudk/gomponents/html"
 )
 
-func TestClasses(t *testing.T) {
-	t.Run("given a map, returns sorted keys from the map with value true", func(t *testing.T) {
-		assert.Equal(t, ` class="boheme-hat hat partyhat"`, Classes{
-			"boheme-hat": true,
-			"hat":        true,
-			"partyhat":   true,
-			"turtlehat":  false,
-		})
-	})
-
-	t.Run("renders as attribute in an element", func(t *testing.T) {
-		e := g.El("div", Classes{"hat": true})
-		assert.Equal(t, `<div class="hat"></div>`, e)
-	})
-
-	t.Run("also works with fmt", func(t *testing.T) {
-		a := Classes{"hat": true}
-		if a.String() != ` class="hat"` {
-			t.FailNow()
-		}
-	})
-}
-
 func TestBooleanAttributes(t *testing.T) {
 	cases := map[string]func() g.Node{
 		"async":     Async,
